backend/requirements.txt
# API Framework
fastapi>=0.97.0,<0.98.0
pydantic>=1.10.0,<2.0.0

# Server
uvicorn>=0.22.0,<0.23.0
python-multipart>=0.0.6,<0.1.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.6,<3.0.0
alembic>=1.11.0,<1.12.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.0.0,<3.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Azure Integration
azure-storage-blob>=12.16.0,<13.0.0
azure-identity>=1.13.0,<2.0.0
azure-core>=1.26.0,<2.0.0

# AI Voice Agent
openai>=1.0.0,<2.0.0
whisper>=1.1.10,<2.0.0
aiofiles>=23.2.1,<24.0.0
librosa>=0.10.1,<0.11.0
soundfile>=0.12.1,<0.13.0

# Development tools
black>=23.3.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
pytest>=7.3.1,<8.0.0
pytest-cov>=4.1.0,<5.0.0
